refactor: 음악 선택 로직 Battle 컴포넌트에 넣기
