+++ src/pages/battle/detail.tsx
@@ -2,12 +2,9 @@ import styled from '@emotion/styled';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import ListIcon from 'public/images/go-to-list-icon.svg';
-import { useState } from 'react';
 
 import Battle from '@/components/battle/detail/Battle/index';
 import { useGetBattle } from '@/components/battle/detail/useGetBattle';
-import { SelectedBattle } from '@/components/battle/types';
-import VoteResult from '@/components/battle/voteResult';
 import BottomNav from '@/components/common/BottomNav';
 import Header from '@/components/common/Header';
 import AuthRequiredPage from '@/components/login/AuthRequiredPage';
@@ -15,18 +12,7 @@ import AuthRequiredPage from '@/components/login/AuthRequiredPage';
 function Detail() {
   const router = useRouter();
   const { id } = router.query;
-
-  const [selectedBattle, setSelectedBattle] = useState<SelectedBattle>({
-    battleId: 0,
-    votedPostId: 0,
-    clickSide: undefined,
-  });
-
   const { data: musicData } = useGetBattle({ initBattleId: Number(id), selectedGenre: 'ALL' });
-
-  const onChangeSelectedBattleInfo = (battleId: number, votedPostId: number, clickSide: 'left' | 'right') => {
-    setSelectedBattle({ battleId, votedPostId, clickSide });
-  };
 
   return id ? (
     <AuthRequiredPage>
@@ -38,17 +24,9 @@ function Detail() {
           </Link>
         }
       />
-      <SelectContainer>
-        <Empty />
-        <Battle musicData={musicData} onChangeSelectedBattleInfo={onChangeSelectedBattleInfo} />
-      </SelectContainer>
-      {selectedBattle.battleId && selectedBattle.votedPostId && (
-        <VoteResult
-          battleId={selectedBattle.battleId}
-          votedPostId={selectedBattle.votedPostId}
-          clickSide={selectedBattle.clickSide}
-        />
-      )}
+      <Container>
+        <Battle musicData={musicData} />
+      </Container>
       <BottomNav />
     </AuthRequiredPage>
   ) : (
@@ -58,13 +36,10 @@ function Detail() {
 
 export default Detail;
 
-const SelectContainer = styled.div`
+const Container = styled.div`
   width: calc(100% - 4rem);
   height: calc(100vh - 16rem);
   min-height: 60rem;
   padding: 0 2rem;
-`;
-
-const Empty = styled.div`
-  height: 1.27rem;
+  padding-top: 1.27rem;
 `;
